+++ client_config.go
@@ -8,3 +8,11 @@ import (
 type ClientConfig struct {
 	http.ClientConfiguration `yaml:",inline"`
 }
+
+// Validate validates the client configuration.
+func (c *ClientConfig) Validate() error {
+	if c.ClientConfiguration.URL == "" {
+		return nil
+	}
+	return c.ClientConfiguration.Validate()
+}