+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.9.6: Added Validate()
+
+This release adds a `Validate()` method to the configuration that allows for check the configuration on loading.
+
 ## 0.9.5: Security configuration
 
 This release adds support for the new [security library](https://github.com/containerssh/security). It also adds a metric for config server requests and updates several libraries to their latest versions.